requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
pyyaml>=6.0
python-multipart>=0.0.6
sse-starlette>=1.8.2
psutil>=5.9.0
httpx>=0.26.0
reportlab>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
aiosqlite>=0.19.0
